refactor(InputView, OutputView): 파일 위치 변경
